Phase 7: Fix analyzer errors - theme, notification, and widget issues
- Fix CardTheme deprecation → CardThemeData
- Fix AppColors constant issues in app_theme.dart
- Remove unused app_colors.dart import
- Fix NotificationService timezone and carryForward issues
- Fix RankingService type casting (num → double)
- Fix TextStyle opacity → Opacity wrapper
- Fix Duration constructor usage
- Add ColorExtensions for withOpacity compatibility
- Improve AppLocalizations.of() error handling

+++ lib/services/ranking_service.dart
@@ -216,7 +216,7 @@ class RankingService {
         totalParticipants: totalParticipants,
         currentUserRank: userRank,
         currentUserScore: userScore,
-        percentile: percentile,
+        percentile: percentile.toDouble(),
         topScore: topScore,
         averageScore: averageScore,
         scoreHistory: scoreHistory,
